Update PartA.py
In this part, at first we used openCv2 and and also iGzFaceToolKit by importing them at the beginning.
Now we could read whole files in a directory and write them to specific file which we want to use them later. whit the named file as "fileList", we try to capture, faces from movies we have, then crop them to save in a specific directory which we want.